requirements.txt
streamlit
ultralytics
Pillow
langchain
faiss-cpu
openai
sentence-transformers
langchain-community 
langchain-groq 
pypdf 
python-dotenv
